+++ km_api/know_me/journal/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'know_me.journal.apps.JournalAppConfig'