+++ km_api/know_me/journal/apps.py
@@ -1,0 +1,8 @@
+from django.apps import AppConfig
+from django.utils.translation import ugettext_lazy as _
+
+
+class JournalAppConfig(AppConfig):
+    label = 'journal'
+    name = 'know_me.journal'
+    verbose_name = _('Know Me - Journal')