+++ km_api/know_me/journal/tests/conftest.py
@@ -1,0 +1,11 @@
+import pytest
+
+from know_me.journal import factories
+
+
+@pytest.fixture
+def entry_factory(db):
+    """
+    Fixture to get the factory used to create journal entries.
+    """
+    return factories.EntryFactory