+++ docs/changelog.rst
@@ -11,6 +11,9 @@ Breaking Changes
   * :issue:`253`: Massive rewrite of profile features. This is a backwards incompatible change that modifies endpoints, the data returned from profile endpoints, and requires a complete database wipe.
   * :issue:`267`: Refactor accessor permission fields. The permissions are now encompassed in a single ``is_admin`` field.
 
+Features
+  * :issue:`246`: Create model for journal entries.
+
 Bug Fixes
   * :issue:`254`: Fix Ansible creating duplicate crontab entries.
   * :issue:`261`: Fix permissions on ``KMUser`` instances not respecting sharing.